fix behaviour of DeleteObjects

+++ src/entrypoint/commands/delete_objects.cpp
@@ -77,15 +77,24 @@ coro<http_response> delete_objects::handle(http_request& req) const {
                                     "xml is invalid");
         }
 
+        bool object_exists;
+        struct object del_object;
+        try {
+            del_object =
+                co_await m_collection.directory.get_object(req.bucket(), *key);
+            object_exists = true;
+        } catch (const command_exception& e) {
+            object_exists = false;
+        }
+
         try {
             LOG_DEBUG() << req.peer() << ": delete_objects::handle(): deleting "
                         << *key;
-
-            auto del_obj =
-                co_await m_collection.directory.get_object(req.bucket(), *key);
             co_await m_collection.directory.delete_object(req.bucket(), *key);
-            co_await m_collection.gdv.unlink(req.context(),
-                                             del_obj.addr.value());
+            if (object_exists) {
+                co_await m_collection.gdv.unlink(req.context(),
+                                                 del_object.addr.value());
+            }
             success.emplace_back(*key);
 
         } catch (const error_exception& e) {
